requirements.txt
python-okx>=0.3.0
pydantic>=2.7
pyyaml>=6.0.1
httpx>=0.27
openai>=1.37
pydantic-settings>=2.4
rich>=13.7
python-dotenv>=1.0
